fix(EEJT-0002) Combining two sort methods into one
